graphql/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
botocore==1.32.5
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.0
gql==3.4.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
py-graphql-client==0.1.2
python-dateutil==2.8.2
python-graphql-client==0.4.3
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.18
websocket-client==0.54.0
websockets==10.4
yarl==1.9.3
